Merge pull request #5 from jakebarn-osu/add-delete-to-customers
From the Step 3 Requirements:
Does the UI have at least one DELETE that will remove things from a M:M relationship? In other words, if an order is deleted from the Orders table, it should also delete the corresponding rows from the OrderDetails table, BUT it should not delete any Products or Customers.

I'm adding a delete to the Customers page so that we can show deleting a Customer updates the M:M Reservations relationship correctly.